Added LibMatrixCuDNN for maintainability and also updated intermediate memory estimates of convolutionop
