+++ src/main/java/org/apache/sysml/hops/ConvolutionOp.java
@@ -223,53 +223,170 @@ public class ConvolutionOp extends Hop  implements MultiThreadedHop
 		return OptimizerUtils.estimateSizeExactSparsity(dim1, dim2, sparsity);
 	}
 	
+	// ---------------------------------------------------------------
+	// Utility methods to guard the computation of memory estimates in presense of unknowns
+	private static class IntermediateDimensions {
+		int dim1; int dim2; double sp;
+		public IntermediateDimensions(ConvolutionOp h, String dim1Str, String dim2Str, double sp) {
+			dim1 = (int) h.getDim(dim1Str);
+			dim2 = (int) h.getDim(dim2Str);
+			this.sp = sp;
+		}
+		public IntermediateDimensions(ConvolutionOp h, String dim1Str, String dim2Str) {
+			dim1 = (int) h.getDim(dim1Str);
+			dim2 = (int) h.getDim(dim2Str);
+			sp = 1;
+		}
+		public IntermediateDimensions(ConvolutionOp h, int dim1, String dim2Str) {
+			this.dim1 = dim1;
+			dim2 = (int) h.getDim(dim2Str);
+			sp = 1;
+		}
+		
+		/**
+		 * Add two computed memory estimates
+		 * 
+		 * @param val1 memory estimate 1
+		 * @param val2 memory estimate 2
+		 * @return sum of memory estimates
+		 */
+		static double add(double val1, double val2) {
+			if(val1 < 0 || val2 < 0) return OptimizerUtils.DEFAULT_SIZE;
+			double ret = val1 + val2;
+			if(ret >= OptimizerUtils.DEFAULT_SIZE) return OptimizerUtils.DEFAULT_SIZE;
+			else return ret;
+		}
+		
+		/**
+		 * Compute memory estimates for given intermediate matrices 
+		 * 
+		 * @param intermediates list of intermediates
+		 * @param numWorkers number of workers
+		 * @return memory estimate
+		 */
+		public static double addEstimateSizes(ArrayList<IntermediateDimensions> intermediates, int numWorkers) {
+			double memBudget = 0; 
+			for(int i = 0; i < intermediates.size(); i++) {
+				memBudget = add(memBudget, OptimizerUtils.estimateSizeExactSparsity(
+						intermediates.get(i).dim1, intermediates.get(i).dim2, intermediates.get(i).sp)*numWorkers);
+			}
+			return memBudget;
+		}
+		
+		/**
+		 * Compute max of two computed memory estimates
+		 * @param val1 memory estimate 1
+		 * @param val2 memory estimate 2
+		 * @return max of memory estimates
+		 */
+		public static double max(double val1, double val2) {
+			if(val1 < 0 || val2 < 0) return OptimizerUtils.DEFAULT_SIZE;
+			double ret = Math.max(val1, val2);
+			if(ret >= OptimizerUtils.DEFAULT_SIZE) return OptimizerUtils.DEFAULT_SIZE;
+			else return ret;
+		}
+	}
+	
+	/**
+	 * Helper utility to compute intermediate memory estimate
+	 * 
+	 * @param gpuIntermediates intermediates for GPU
+	 * @param cpIntermediates intermediates for CP
+	 * @return memory estimates
+	 */
+	private double computeIntermediateMemEstimateHelper(
+			ArrayList<IntermediateDimensions> gpuIntermediates,
+			ArrayList<IntermediateDimensions> cpIntermediates) {
+		// Since CP operators use row-level parallelism by default
+		int numWorkers = (int) Math.min(OptimizerUtils.getConstrainedNumThreads(_maxNumThreads), Math.max(getDim("N"), 1));
+		if(DMLScript.USE_ACCELERATOR) {
+			// Account for potential sparse-to-dense conversion
+			double gpuMemBudget = IntermediateDimensions.addEstimateSizes(gpuIntermediates, 1);
+			double cpMemoryBudget = IntermediateDimensions.addEstimateSizes(cpIntermediates, numWorkers);
+			if(cpMemoryBudget > gpuMemBudget) {
+				double oneThreadCPMemBudget = IntermediateDimensions.addEstimateSizes(cpIntermediates, 1);
+				if(oneThreadCPMemBudget <= gpuMemBudget) {
+					// Why limit CPU ? in-order to give more opportunity to compile GPU operators
+					cpMemoryBudget = oneThreadCPMemBudget;
+				}
+			}
+			// Finally, use the maximum of CP and GPU memory budget
+			return IntermediateDimensions.max(cpMemoryBudget, gpuMemBudget);
+		}
+		else {
+			// When -gpu flag is not provided, the memory estimates for CP are not affected.
+			return IntermediateDimensions.addEstimateSizes(cpIntermediates, numWorkers);
+		}
+	}
+	
 	@Override
 	protected double computeIntermediateMemEstimate( long dim1, long dim2, long nnz )
 	{	
-		double ret = 0;
-		ConvolutionParameters params;
-		try {
-			params = parseInput();
-		} catch (DMLRuntimeException e) {
-			throw new RuntimeException(e);
-		}
-		
-		// Since CP operators use row-level parallelism by default
-		int numWorkers = Math.min(OptimizerUtils.getConstrainedNumThreads(_maxNumThreads), Math.max(params.N, 1));
-		
+		ArrayList<IntermediateDimensions> gpuIntermediates = new ArrayList<IntermediateDimensions>();
+		ArrayList<IntermediateDimensions> cpIntermediates = new ArrayList<IntermediateDimensions>();
 		if(getOp() == ConvOp.DIRECT_CONV2D) {
-			double inputSparsity = getInput().get(0).getSparsity();
-			// Why inputSparsity => im2col operation preserves the worst-case sparsity of the input.
-			double cpMemBudget = OptimizerUtils.estimateSizeExactSparsity(params.C*params.R*params.S, params.P*params.Q*numWorkers, inputSparsity);
-			if(DMLScript.USE_ACCELERATOR) {
-				// Assumption: To compile a GPU conv2d operator, following should fit on the GPU:
-				// 1. output in dense format (i.e. computeOutputMemEstimate) 
-				// 2. input in any format
-				// 3. atleast one input row in dense format
-				// 4. filter in dense format
-				
-				// Account for potential sparse-to-dense conversion of atleast 1 input row and filter
-				double gpuMemBudget = OptimizerUtils.estimateSizeExactSparsity(1, dim2, 1.0) + OptimizerUtils.estimateSizeExactSparsity(params.K, params.C*params.R*params.S, 1.0);
-				
-				if(cpMemBudget > gpuMemBudget) {
-					// Limit the CP memory budget to allow for more opportunity for compiling GPU instructions
-					// Constraint: allow for atleast one im2col intermediate in case the optimizer decides to compile CP instruction
-					cpMemBudget = OptimizerUtils.estimateSizeExactSparsity(params.C*params.R*params.S, params.P*params.Q, inputSparsity);
-				}
-				
-				// Finally, use the maximum of CP and GPU memory budget
-				ret += Math.max(cpMemBudget, gpuMemBudget);
-			}
-			else {
-				// When -gpu flag is not provided, the memory estimates for CP are not affected.
-				ret += cpMemBudget;
-			}
+			// Assumption: To compile a GPU conv2d operator, following should fit on the GPU:
+			// 1. output in dense format (i.e. computeOutputMemEstimate) 
+			// 2. input in any format
+			// 3. atleast one input row in dense format
+			// 4. filter in dense format
+			
+			// Account for potential sparse-to-dense conversion of atleast 1 input row and filter
+			gpuIntermediates.add(new IntermediateDimensions(this, 1, "CHW"));
+			gpuIntermediates.add(new IntermediateDimensions(this, "K", "CRS"));
+			
+			// im2col operation preserves the worst-case sparsity of the input.
+			cpIntermediates.add(new IntermediateDimensions(this, "CRS", "PQ", getInput().get(0).getSparsity()));
+		}
+		else if(getOp() == ConvOp.DIRECT_CONV2D_BACKWARD_DATA) {
+			// Assumption: To compile a GPU conv2d_backward_data operator, following should fit on the GPU:
+			// 1. output in dense format (i.e. computeOutputMemEstimate) 
+			// 2. dout in any format
+			// 3. atleast one dout row in dense format
+			// 4. filter in dense format
+			
+			// Account for potential sparse-to-dense conversion of atleast 1 input row and filter
+			gpuIntermediates.add(new IntermediateDimensions(this, 1, "KPQ"));
+			gpuIntermediates.add(new IntermediateDimensions(this, "K", "CRS"));
+			
+			// There are 2 intermediates: rotate180 and input to col2im for conv2d_backward_data
+			// rotate180 preserves the "exact" sparsity of the dout matrix
+			cpIntermediates.add(new IntermediateDimensions(this, "PQ", "K", getInput().get(1).getSparsity()));
+			// Note: worst-case sparsity for the input of col2im (of size NPQ x CRS where N is determined by degree of parallelism)
+			cpIntermediates.add(new IntermediateDimensions(this, "PQ", "CRS"));
+		}
+		else if(getOp() == ConvOp.DIRECT_CONV2D_BACKWARD_FILTER) {
+			// Assumption: To compile a GPU conv2d_backward_filter operator, following should fit on the GPU:
+			// 1. output in dense format (i.e. computeOutputMemEstimate) 
+			// 2. dout in any format
+			// 3. atleast one dout and input row in dense format
+			
+			// Account for potential sparse-to-dense conversion of atleast 1 input + dout row
+			gpuIntermediates.add(new IntermediateDimensions(this, 1, "CHW"));
+			gpuIntermediates.add(new IntermediateDimensions(this, 1, "KPQ"));
+			
+			// There are 2 intermediates: im2col and rotate180 for conv2d_backward_filter
+			// rotate180 preserves the "exact" sparsity of the dout matrix
+			cpIntermediates.add(new IntermediateDimensions(this, "PQ", "K", getInput().get(1).getSparsity()));
+			// im2col operation preserves the worst-case sparsity of the input.
+			cpIntermediates.add(new IntermediateDimensions(this, "CRS", "PQ", getInput().get(0).getSparsity()));
+		}
+		else if(getOp() == ConvOp.MAX_POOLING) {
+			// Account for potential sparse-to-dense conversion of atleast 1 input row
+			gpuIntermediates.add(new IntermediateDimensions(this, 1, "CHW"));
+		}
+		else if(getOp() == ConvOp.MAX_POOLING_BACKWARD) {
+			// Account for potential sparse-to-dense conversion of atleast 1 input + dout row
+			gpuIntermediates.add(new IntermediateDimensions(this, 1, "CHW"));
+			gpuIntermediates.add(new IntermediateDimensions(this, 1, "CPQ"));
 		}
 		
-		
-		
-		return ret;
+		if(gpuIntermediates.size() > 0 || cpIntermediates.size() > 0)
+			return computeIntermediateMemEstimateHelper(gpuIntermediates, cpIntermediates);
+		else
+			return 0;
 	}
+	
 	
 	@Override
 	protected long[] inferOutputCharacteristics( MemoTable memo )
@@ -284,64 +401,41 @@ public class ConvolutionOp extends Hop  implements MultiThreadedHop
 			ret[2] = -1;
 			return (ret[0]>0 && ret[1]>0) ? ret : null;
 		}
-	
-		ConvolutionParameters params;
-		try {
-			params = parseInput();
-		} catch (DMLRuntimeException e) {
-			throw new RuntimeException(e);
-		}
 		
 		switch(op) 
 		{
 			case MAX_POOLING: {
-				// input
-				long N = getInput().get(0)._dim1;
-				ret[0] = N;
-				ret[1] = getExtractedVal(params.C, params.P, params.Q);
+				ret[0] = getDim("N");
+				ret[1] = getDim("CPQ");
 				ret[2] = -1;
 				break;
 			}
 			case DIRECT_CONV2D: {
-				// input, filter
-				long N = getInput().get(0)._dim1;
-				ret[0] = N;
-				ret[1] = getExtractedVal(params.K, params.P, params.Q);
+				ret[0] = getDim("N");
+				ret[1] = getDim("KPQ");
 				ret[2] = -1;
 				break;
 			}
 			case DIRECT_CONV2D_BACKWARD_FILTER: {
-				// input, dout	
-				ret[0] = params.K;
-				ret[1] = getExtractedVal(params.C, params.R, params.S);
+				ret[0] = getDim("K");
+				ret[1] = getDim("CRS");
 				ret[2] = -1;
 				break;
 			}
 			case MAX_POOLING_BACKWARD: {
-				// input, dout
-				ret[0] = getInput().get(0)._dim1;
-				ret[1] = getInput().get(0)._dim2;
+				ret[0] = getDim("N");
+				ret[1] = getDim("CHW");
 				ret[2] = -1;
 				break;
 			}
 			case DIRECT_CONV2D_BACKWARD_DATA: {
-				// filter, dout
-				long N = getInput().get(1)._dim1;
-				ret[0] = N;
-				ret[1] = getExtractedVal(params.C, params.H, params.W);
+				ret[0] = getDim("N");
+				ret[1] = getDim("CHW");
 				ret[2] = -1;
 				break;
 			}
 			default:
 				throw new RuntimeException("Unsupported op:" + op.name());
-		}
-		
-		if(LOG.isDebugEnabled() && (ret[0] <= 0 || ret[1] <= 0)) {
-			LOG.debug("Unknown dimensions for ConvolutionOp in inferOutputCharacteristics:" + op.name() + " " + ret[0] + " " + ret[1] + 
-					" img_dim=[" + params.N + " " + params.C + " " + params.H + " " + params.W + "]" +
-					" filter_dim=[" + params.K + " " + params.C + " " + params.H + " " + params.W + "]" + 
-					" output_feature_map=[" + params.P + " " + params.Q + "] stride=[" + params.stride_h + " " + params.stride_w + "]" +
-					" pad=[" + params.pad_h + " " + params.pad_w + "]");
 		}
 		
 		//safe return (create entry only if at least dims known)
@@ -427,13 +521,6 @@ public class ConvolutionOp extends Hop  implements MultiThreadedHop
 		}
 		return _cachedParams;
 	}
-
-	public static long getExtractedVal(long val1, long val2, long val3) {
-		if(val1 == -1 || val2 == -1 || val3 == -1) {
-			return -1;
-		}
-		return val1*val2*val3;
-	}
 	
 	@Override
 	public void refreshSizeInformation()
@@ -442,71 +529,49 @@ public class ConvolutionOp extends Hop  implements MultiThreadedHop
 			Hop input1 = getInput().get(0);
 			setDim1(input1.getDim1());
 			setDim2(input1.getDim2());
+			_nnz = -1; // cannot infer stats
 			return;
-		}
-		
-		ConvolutionParameters params;
-		try {
-			params = parseInput();
-		} catch (DMLRuntimeException e) {
-			throw new RuntimeException(e);
 		}
 		
 		switch(op) 
 		{
 			case MAX_POOLING:
 			{	
-				// input
-				long N = getInput().get(0)._dim1;
-				_dim1 = N;
-				_dim2 = getExtractedVal(params.C, params.P, params.Q);
+				_dim1 = getDim("N");
+				_dim2 = getDim("CPQ");
 				_nnz = -1; // cannot infer stats
 				break;
 			}
 			case MAX_POOLING_BACKWARD:
 			{
-				// input, dout
-				_dim1 = getInput().get(0)._dim1;
-				_dim2 = getInput().get(0)._dim2;
+				_dim1 = getDim("N");
+				_dim2 = getDim("CHW");
 				_nnz = -1;
 				break;
 			}
 			case DIRECT_CONV2D:
 			{
-				// input, filter
-				long N = getInput().get(0)._dim1;
-				_dim1 = N;
-				_dim2 = getExtractedVal(params.K, params.P, params.Q);
+				_dim1 = getDim("N");
+				_dim2 = getDim("KPQ");
 				_nnz = -1; // cannot infer stats
 				break;
 			}
 			case DIRECT_CONV2D_BACKWARD_DATA:
 			{
-				// filter, dout
-				long N = getInput().get(1)._dim1;
-				_dim1 = N;
-				_dim2 = getExtractedVal(params.C, params.H, params.W);
+				_dim1 = getDim("N");
+				_dim2 = getDim("CHW");
 				_nnz = -1; // cannot infer stats
 				break;
 			}
 			case DIRECT_CONV2D_BACKWARD_FILTER:
 			{
-				// input, dout	
-				_dim1 = params.K;
-				_dim2 = getExtractedVal(params.C, params.R, params.S);
+				_dim1 = getDim("K");
+				_dim2 = getDim("CRS");
 				_nnz = -1; // cannot infer stats
 				break;
 			}
 			default:
 				throw new RuntimeException("The sizes are not refreshed for " + op.name());
-		}
-		
-		if(LOG.isDebugEnabled() && (_dim1 <= 0 || _dim2 <= 0)) {
-			LOG.debug("Unknown dimensions for ConvolutionOp in refreshSizeInformation:" + op.name() + " " + _dim1 + " " + _dim2 + 
-					" img_dim=[" + params.N + " " + params.C + " " + params.H + " " + params.W + "]" +
-					" filter_dim=[" + params.K + " " + params.C + " " + params.H + " " + params.W + "]" + 
-					" output_feature_map=[" + params.P + " " + params.Q + "] stride=[" + params.stride_h + " " + params.stride_w + "]" +
-					" pad=[" + params.pad_h + " " + params.pad_w + "]");
 		}
 	}
 	
@@ -553,4 +618,132 @@ public class ConvolutionOp extends Hop  implements MultiThreadedHop
 	public int getMaxNumThreads() {
 		return _maxNumThreads;
 	}
+	
+	
+	// ------------------------------------------------------------------------------------------------------
+	// Utility methods to get the dimensions taking into account unknown dimensions
+	
+	/**
+	 * Convenient method to get the dimensions required by ConvolutionOp.
+	 * 
+	 * @param dimString can be K, CRS, N, CHW, KPQ, PQ
+	 * @return either -1 or value associated with the dimString
+	 */
+	private long getDim(String dimString) {
+		if(op == ConvOp.BIAS_ADD || op == ConvOp.BIAS_MULTIPLY) {
+			throw new RuntimeException("getDim method should not be invoked for bias_add and bias_multiply");
+		}
+		ConvolutionParameters params;
+		try {
+			params = parseInput();
+		} catch (DMLRuntimeException e) {
+			throw new RuntimeException(e);
+		}
+		Hop filter = null; 	// shape: K x CRS 
+		Hop input = null; 	// shape: N x CHW
+		Hop dout = null;	// shape: N x KPQ
+		Hop dout1 = null;	// shape: N x CPQ
+		
+		if(getOp() == ConvOp.DIRECT_CONV2D) {
+			input  = getInput().get(0);
+			filter = getInput().get(1);
+		}
+		else if(getOp() == ConvOp.DIRECT_CONV2D_BACKWARD_DATA) {
+			filter = getInput().get(0);
+			dout  = getInput().get(1);
+		}
+		else if(getOp() == ConvOp.DIRECT_CONV2D_BACKWARD_FILTER) {
+			input = getInput().get(0);
+			dout  = getInput().get(1);
+		}
+		else if(getOp() == ConvOp.MAX_POOLING) {
+			input = getInput().get(0);
+		}
+		else if(getOp() == ConvOp.MAX_POOLING_BACKWARD) {
+			input = getInput().get(0);
+			dout1  = getInput().get(1);
+		}
+		
+		long ret = -1;
+		if(dimString.equals("K") && filter != null) {
+			ret = getNonNegative(params.K, filter._dim1);
+		}
+		else if(dimString.equals("CRS") && filter != null) {
+			ret = getNonNegative(nonNegativeMultiply(params.C, params.R, params.S), filter._dim2);
+		}
+		else if(dimString.equals("N") && input != null) {
+			ret = getNonNegative(params.N, input._dim1);
+		}
+		else if(dimString.equals("CHW") && input != null) {
+			ret = getNonNegative(nonNegativeMultiply(params.C, params.H, params.W), input._dim2);
+		}
+		else if(dimString.equals("N") && dout != null) {
+			ret = getNonNegative(params.N, dout._dim1);
+		}
+		else if(dimString.equals("KPQ") && dout != null) {
+			ret = getNonNegative(nonNegativeMultiply(params.K, params.P, params.Q), dout._dim2);
+		}
+		else if(dimString.equals("N") && dout1 != null) {
+			ret = getNonNegative(params.N, dout1._dim1);
+		}
+		else if(dimString.equals("CPQ") && dout1 != null) {
+			ret = getNonNegative(nonNegativeMultiply(params.C, params.P, params.Q), dout1._dim2);
+		}
+		else if(dimString.equals("K")) {
+			ret = params.K >= 0 ? params.K : -1;
+		}
+		else if(dimString.equals("CRS")) {
+			ret = nonNegativeMultiply(params.C, params.R, params.S);
+		}
+		else if(dimString.equals("N")) {
+			ret = params.N >= 0 ? params.N : -1;
+		}
+		else if(dimString.equals("CHW")) {
+			ret = nonNegativeMultiply(params.C, params.H, params.W);
+		}
+		else if(dimString.equals("KPQ")) {
+			ret = nonNegativeMultiply(params.K, params.P, params.Q);
+		}
+		else if(dimString.equals("PQ")) {
+			ret = nonNegativeMultiply(params.P, params.Q);
+		}
+		else if(dimString.equals("CPQ")) {
+			ret = nonNegativeMultiply(params.C, params.P, params.Q);
+		}
+		else {
+			throw new RuntimeException("Unsupported dimension:" + dimString + " for operator " + getOp().name());
+		}
+		
+		if(LOG.isDebugEnabled() && ret <= 0) {
+			LOG.debug("Unknown dimension " + dimString + " for ConvolutionOp:" + op.name() + 
+					" img_dim=[" + params.N + " " + params.C + " " + params.H + " " + params.W + "]" +
+					" filter_dim=[" + params.K + " " + params.C + " " + params.H + " " + params.W + "]" + 
+					" output_feature_map=[" + params.P + " " + params.Q + "] stride=[" + params.stride_h + " " + params.stride_w + "]" +
+					" pad=[" + params.pad_h + " " + params.pad_w + "]");
+		}
+		return ret;
+	}
+	
+	private long nonNegativeMultiply(long val1, long val2, long val3) {
+		if(val1 >= 0 && val2 >= 0 && val3 >= 0) {
+			return val1 * val2 * val3;
+		}
+		else return -1;
+	}
+	private long nonNegativeMultiply(long val1, long val2) {
+		if(val1 >= 0 && val2 >= 0) {
+			return val1 * val2;
+		}
+		else return -1;
+	}
+	private long getNonNegative(long val1, long val2) {
+		if(val1 >= 0 && val2 >= 0) {
+			if(val1 == val2) return val1;
+			else throw new RuntimeException("Incorrect dimensions in Convolution Hop: " + val1 + " != " + val2);
+		}
+		else if(val1 >= 0) return val1;
+		else if(val2 >= 0) return val2;
+		else return -1;
+	}
+	// ------------------------------------------------------------------------------------------------------
 }
